das formatter: special tokens and ignore file feature
Since we want to add formatting validation to das scripts, we also need a mechanism that allows us to quickly exclude individual files from formatting. In this regard, it was proposed to add special formatting tokens starting with a prefix "//fmt:". The first such token is "//fmt:ignore-file" which can disable the formatting of the current file.