worker contexts feature
This commit implements a functionality of daScript "worker" contexts.
The main task of these contexts is to perform part of the work in worker threads for the daScript entity component system queries(One worker context per one worker thread).
These contexts can be created in advance, can have their own local stack and heap, and can share global constants with origin context(read only).
So during the MT execution, all we need is to choose the worker context corresponding to the current worker thread and execute for it makeWorkerFor method by passing there a pointer to the origin context. In the main thread, we just use the origin context.