Fix block invoke AOT codegen not compiling smtimes
One of the possible forms of invoking a block from AOT
codegen code is das_invoke<ResType>::invoke<ArgTypes...>
(or same with invoke_cmres). What we actually pass to this
function can be different depending on the context: an actual
Block object, OR a **C++ lambda**.

When passing a lambda, we have no choice but to use template
argument deduction. But that was completely broken, as the
lambda type BLK was specified as the first template argument
for invoke/invoke_cmres and so AOT codegen tried to specify
the lambda type as some argument type instead.

Moving the BLK template parameter to be the last one seems
to fix the problem. BUT! It brings a new problem, now the
templated overload becomes preferable over regular overloads
when upcasting the block to a parent type.

To fight the new issue, a cheeky enable_if was employed.