doc: support sphinx versions above 7.3.0
sphinx rearranged code for python domain splitting it into smaller
parts, so versions newer or equal 7.3.0 have this function in different
place. Dagor has moved to sphinx 8+ for documentation and we need this
change.

Sphinx8 has mostly deprecated string-like paths and is now (almost
everywhere) using PathLike objects for which addition with strings is
not possible, so f-strings are now required