jit: fix inscope use no init
Code in inscope_return_inscope.das contained UB, code like:
```
[sideeffects]
def buggy(skip : bool) {
    var a = new Foo
    if (skip) {
        return
    }
    var b = new Foo
} finally {
    unsafe {
        delete a
        delete b
    }
}
```
Sometimes it crashed, because b is uninitialized and when delete
was called it was treated as initialized.

More specifically it crashed when `daScriptTest` was called with
all the JIT optimizations disabled (LLVM_ENABLE_OPT_PASS=false)
and daslang built in Release or RelWithDebInfo.